fix: 将颜色常量从 uint32_t 改为 uint16_t 以正确解析 RGB565
根因：LovyanGFX 的 color_conv_t::convert(uint32_t) 将 uint32_t
颜色值当作 RGB888 处理，而 convert_to_rgb888(uint32_t) 直接返回
c & 0xFFFFFF。代码中定义的 RGB565 值（如 0x10A2）被定义为
uint32_t，导致被错误解析为 RGB888：
- 0x10A2 → RGB(0, 16, 162) = 深蓝色（应为深灰色）
- 0x2104 → RGB(0, 33, 4) = 深绿色（应为深灰色）
- 0x2124 → RGB(0, 33, 36) = 青色（应为浅灰点）
- 0x03DF → RGB(0, 3, 223) = 亮蓝色（与预期接近但有偏差）

这导致整个 UI 背景、header、dots 颜色全部错乱，形成色块和断裂感。

修改（src/UIScreen.cpp）：
- 所有自定义颜色常量从 uint32_t 改为 uint16_t
- M5GFX 对 uint16_t 会正确调用 convert_rgb565，按 RGB565 解析

影响范围：所有屏幕的背景色、点阵、高亮色、文本颜色

+++ src/UIScreen.cpp
@@ -1,11 +1,12 @@
 #include "UIScreen.h"
 
 // Define modern colors
-static constexpr uint32_t COLOR_BG = 0x10A2; // RGB565(16,16,16) roughly
-static constexpr uint32_t COLOR_DOT = 0x2124; // RGB565(32,32,32)
-static constexpr uint32_t COLOR_ACCENT = 0x03DF; // Bright Blue matching the button (RGB: 0, 122, 255)
-static constexpr uint32_t COLOR_TEXT = TFT_WHITE;
-static constexpr uint32_t COLOR_HEADER = 0x2104; // Darker grey
+// IMPORTANT: Use uint16_t for RGB565 values. uint32_t is interpreted as RGB888 by M5GFX.
+static constexpr uint16_t COLOR_BG = 0x10A2; // RGB565(16,16,16) roughly
+static constexpr uint16_t COLOR_DOT = 0x2124; // RGB565(32,32,32)
+static constexpr uint16_t COLOR_ACCENT = 0x03DF; // Bright Blue matching the button (RGB: 0, 122, 255)
+static constexpr uint16_t COLOR_TEXT = TFT_WHITE;
+static constexpr uint16_t COLOR_HEADER = 0x2104; // Darker grey
 
 void UIScreen::init() {
   M5.Display.setRotation(0);  // 135x240
